requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.1.0
braceexpand==0.1.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
discord.py==2.0.1
frozenlist==1.3.3
idna==3.4
import-expression==1.1.4
jishaku==2.5.1
line-profiler==3.5.1
multidict==6.0.2
python-decouple==3.6
six==1.16.0
yarl==1.8.1
